requirements.txt
accelerate==1.7.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
diffusers==0.33.1
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
hf-xet==1.1.2
httplib2==0.22.0
huggingface-hub==0.32.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
numpy
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydub==0.25.1
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
sentencepiece==0.2.0
sympy==1.14.0
tokenizers==0.21.1
torch
torchvision
torchaudio
transformers==4.52.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
zipp==3.21.0
pyaudio
